refactor: modularize data, state, and engine
