+++ data/realms.js
@@ -1,0 +1,12 @@
+export const REALMS = [
+  {name:'Mortal', stages:9, cap:100, fcap:60, baseRegen:1, atk:1, def:1, bt:0.60, power:1},
+  {name:'Qi Refining', stages:9, cap:300, fcap:150, baseRegen:2, atk:3, def:2, bt:0.55, power:3},
+  {name:'Foundation', stages:9, cap:800, fcap:400, baseRegen:3, atk:6, def:4, bt:0.50, power:8},
+  {name:'Core', stages:9, cap:2000, fcap:1100, baseRegen:5, atk:12, def:8, bt:0.45, power:20},
+  {name:'Nascent', stages:9, cap:7000, fcap:3000, baseRegen:8, atk:24, def:16, bt:0.40, power:50},
+  {name:'Soul Formation', stages:9, cap:20000, fcap:8000, baseRegen:12, atk:40, def:28, bt:0.35, power:120},
+  {name:'Void Refining', stages:9, cap:60000, fcap:25000, baseRegen:18, atk:70, def:50, bt:0.30, power:300},
+  {name:'Body Integration', stages:9, cap:180000, fcap:80000, baseRegen:25, atk:120, def:85, bt:0.25, power:750},
+  {name:'Mahayana', stages:9, cap:500000, fcap:250000, baseRegen:35, atk:200, def:140, bt:0.20, power:1800},
+  {name:'Tribulation', stages:9, cap:1500000, fcap:800000, baseRegen:50, atk:350, def:250, bt:0.15, power:4500}
+];
