+++ index.html
@@ -748,6 +748,6 @@
     </section>
   </main>
 
-  <script type="module" src="game.js"></script>
+  <script type="module" src="ui/index.js"></script>
 </body>
 </html>
